run tests on LUT

+++ src/test/java/com/example/SampleClientTest.java
@@ -1,6 +1,7 @@
 package com.example;
 
 import com.example.product.invoice.CombinedInvoiceDocument;
+import com.mindee.http.MindeeHttpException;
 import com.mindee.input.LocalInputSource;
 import org.junit.jupiter.api.Test;
 import java.io.File;
@@ -18,24 +19,35 @@ public class SampleClientTest {
   }
 
   protected void runFile(File filePath) throws IOException, InterruptedException {
+    // generate an output path
+    Path output = Paths.get(filePath.getPath().replace(".pdf", ".rst"));
+    if (Files.exists(output)) {
+      System.out.println("File exists: " + output);
+      return;
+    }
+
     LocalInputSource inputSource = new LocalInputSource(filePath);
-    CombinedInvoiceDocument invoice = client.parseLvaInvoice(inputSource);
+    CombinedInvoiceDocument invoice = client.parseLutInvoice(inputSource);
 
     // print to console the invoice data as rST
     System.out.println(invoice);
 
     // save the contents in a rST file
-    Path output = Paths.get(filePath.getPath().replace(".pdf", ".rst"));
     byte[] bytes = invoice.toString().getBytes(StandardCharsets.UTF_8);
     Files.write(output, bytes);
   }
 
   @Test
   public void runFiles() throws IOException, InterruptedException {
-    File folder = new File("/home/ianare/Documents/lva");
+    File folder = new File("/home/ianare/Documents/lut");
     File[] files = folder.listFiles((dir, name) -> name.endsWith(".pdf"));
     for (File file : files) {
-      runFile(file);
+      try {
+        runFile(file);
+      } catch (MindeeHttpException err) {
+        System.out.println(file.getPath());
+        System.out.println(err.toString());
+      }
     }
   }
 }
